+++ src/main/java/carrace/ui/InputView.java
@@ -1,4 +1,4 @@
-package carrace;
+package carrace.ui;
 
 import java.util.Scanner;
