+++ src/test/java/carrace/domain/RaceTrackTest.java
@@ -1,4 +1,4 @@
-package carrace;
+package carrace.domain;
 
 import org.junit.jupiter.api.RepeatedTest;
